--- src/client/src/features/common/aspectobject/index.ts
@@ -1,1 +1,0 @@
-export { AspectObjectPreview } from "features/common/aspectobject/AspectObjectPreview";
