+++ src/client/src/features/entities/entityPreviews/aspectobject/AspectObjectPreview.helpers.ts
