+++ src/client/src/features/entities/entityPreviews/attribute/AttributeFormPreview.tsx
@@ -1,6 +1,6 @@
 import { Control, useWatch } from "react-hook-form";
 import { useTranslation } from "react-i18next";
-import { FormAttributeLib } from "../attributes/types/formAttributeLib";
+import { FormAttributeLib } from "../../attributes/types/formAttributeLib";
 import AttributePreview from "./AttributePreview";
 
 interface AttributeFormPreviewProps {
