+++ src/client/src/features/entities/entityPreviews/terminal/TerminalFormPreview.tsx
@@ -1,4 +1,4 @@
-import { TerminalPreview } from "features/entities/entityPreviews/TerminalPreview";
+import { TerminalPreview } from "features/entities/entityPreviews/terminal/TerminalPreview";
 import { FormTerminalLib } from "features/entities/terminal/types/formTerminalLib";
 import { Control, useWatch } from "react-hook-form";
 import { useTranslation } from "react-i18next";
