+++ src/client/src/features/entities/entityPreviews/terminal/TerminalPreview.styled.tsx
