+++ src/client/src/features/explore/about/About.tsx
@@ -10,11 +10,11 @@ import { SelectedInfo } from "features/explore/common/selectedInfo";
 import { useEffect, useState } from "react";
 import { useTranslation } from "react-i18next";
 import { useGetAttribute } from "../../../external/sources/attribute/attribute.queries";
-import AttributePreview from "../../entities/entityPreviews/AttributePreview";
+import AttributePreview from "../../entities/entityPreviews/attribute/AttributePreview";
 import { toFormAttributeLib } from "../../entities/attributes/types/formAttributeLib";
-import UnitPreview from "../../entities/entityPreviews/UnitPreview";
-import DatumPreview from "../../entities/entityPreviews/DatumPreview";
-import { RdsPreview } from "../../entities/entityPreviews/RdsPreview";
+import UnitPreview from "../../entities/entityPreviews/unit/UnitPreview";
+import DatumPreview from "../../entities/entityPreviews/datum/DatumPreview";
+import { RdsPreview } from "../../entities/entityPreviews/rds/RdsPreview";
 import { toFormUnitLib } from "../../entities/units/types/formUnitLib";
 import { useGetUnit } from "../../../external/sources/unit/unit.queries";
 import { useGetDatum } from "../../../external/sources/datum/datum.queries";
