+++ src/client/src/features/explore/about/components/terminal/TerminalPanel.tsx
@@ -3,7 +3,7 @@ import { Token } from "complib/general";
 import { Flexbox, MotionBox } from "complib/layouts";
 import { Heading, Text } from "complib/text";
 import { InfoItemButton } from "features/common/info-item";
-import { TerminalPreview } from "features/entities/entityPreviews/TerminalPreview";
+import { TerminalPreview } from "features/entities/entityPreviews/terminal/TerminalPreview";
 import { PanelPropertiesContainer } from "features/explore/about/components/common/PanelPropertiesContainer";
 import { PanelSection } from "features/explore/about/components/common/PanelSection";
 import { useTranslation } from "react-i18next";
